+++ packages/dnb-design-system-portal/src/shared/tags/Anchor.js
@@ -5,16 +5,77 @@
 
 import React from 'react'
 import PropTypes from 'prop-types'
-import classnames from 'classnames'
-import AnchorLink from 'react-anchor-link-smooth-scroll'
+import { Link } from '@dnb/eufemia/src/elements'
 
-const Anchor = ({ children, className, href, ...rest }) => {
+export class AnchorLink extends React.PureComponent {
+  static propTypes = {
+    element: PropTypes.oneOfType([PropTypes.node, PropTypes.object]),
+    onClick: PropTypes.func,
+    offset: PropTypes.string
+  }
+
+  static defaultProps = {
+    element: 'a',
+    onClick: null,
+    offset: null
+  }
+
+  onClickHandler = (e) => {
+    if (this.props.onClick) {
+      e.preventDefault()
+
+      let offset = () => 0
+      if (this.props.offset !== null) {
+        if (
+          this.props.offset &&
+          this.props.offset.constructor &&
+          this.props.offset.apply
+        ) {
+          offset = this.props.offset
+        } else {
+          offset = () => parseInt(this.props.offset)
+        }
+      }
+
+      const id = e.currentTarget.getAttribute('href').slice(1)
+      const anchorElem = document.getElementById(id)
+      const offsetTop =
+        (anchorElem &&
+          anchorElem.getBoundingClientRect().top + window.pageYOffset) ||
+        0
+
+      window.scroll({
+        top: offsetTop - offset(),
+        behavior: 'smooth'
+      })
+
+      this.props.onClick(e)
+    }
+  }
+
+  render() {
+    const {
+      element,
+      offset, // eslint-disable-line
+      ...rest
+    } = this.props
+
+    rest.onClick = this.onClickHandler
+    const Element = element || 'a'
+
+    // eslint-disable-next-line jsx-a11y/anchor-has-content
+    return <Element {...rest} />
+  }
+}
+
+const Anchor = ({ children, href, ...rest }) => {
   if (/^#/.test(href)) {
     return (
       <AnchorLink
+        lang="en-GB"
+        element={Link}
         offset="100"
         href={href}
-        className={classnames('dnb-anchor', className)}
         {...rest}
       >
         {children}
@@ -29,22 +90,16 @@ const Anchor = ({ children, className, href, ...rest }) => {
     }
   }
   return (
-    <a
-      href={href}
-      className={classnames('dnb-anchor', className)}
-      {...rest}
-    >
+    <AnchorLink lang="en-GB" element={Link} href={href} {...rest}>
       {children}
-    </a>
+    </AnchorLink>
   )
 }
 Anchor.propTypes = {
   href: PropTypes.string.isRequired,
-  className: PropTypes.string,
   children: PropTypes.node
 }
 Anchor.defaultProps = {
-  className: null,
   children: null
 }