+++ packages/dnb-eufemia/src/components/line-title/index.js
@@ -1,8 +1,0 @@
-/**
- * Component Entry
- *
- */
-
-import LineTitle from './LineTitle'
-export default LineTitle
-export * from './LineTitle'