+++ packages/dnb-eufemia/src/components/notification/index.js
@@ -1,8 +1,0 @@
-/**
- * Component Entry
- *
- */
-
-import Notification from './Notification'
-export default Notification
-export * from './Notification'