+++ packages/dnb-eufemia/src/style/elements/__tests__/Anchor.screenshot.test.js
@@ -14,7 +14,7 @@ describe('Anchor screenshot', () => {
     url: '/uilib/elements/anchor',
     screenshotConfig: {
       // use 6% on CI because of the font rendering differences
-      pixelThresholdRelative: isCI ? 0.06 : 0
+      pixelThresholdRelative: isCI ? 0.07 : 0
     }
   })
   // the first one is on 5.54%
@@ -26,7 +26,7 @@ describe('Anchor screenshot', () => {
   })
   it('have to match the target blank state', async () => {
     const screenshot = await testPageScreenshot({
-      selector: '[data-visual-test="anchor-blank"]'
+      selector: '[data-visual-test="anchor-blank"] a'
     })
     expect(screenshot).toMatchImageSnapshot()
   })
@@ -54,13 +54,6 @@ describe('Anchor screenshot', () => {
     })
     expect(screenshot).toMatchImageSnapshot()
   })
-  it('have to match the target blank on active state', async () => {
-    const screenshot = await testPageScreenshot({
-      selector: '[data-visual-test="anchor-blank"]',
-      simulate: 'active'
-    })
-    expect(screenshot).toMatchImageSnapshot()
-  })
   it('have to match the "hover" state', async () => {
     const screenshot = await testPageScreenshot({
       selector: '[data-visual-test="anchor-hover"]'
@@ -77,6 +70,18 @@ describe('Anchor screenshot', () => {
     const screenshot = await testPageScreenshot({
       selector: '[data-visual-test="anchor-focus"]',
       simulate: 'focus' // should be tested first
+    })
+    expect(screenshot).toMatchImageSnapshot()
+  })
+  it('have to match the target blank on active state', async () => {
+    const screenshot = await testPageScreenshot({
+      style: {
+        'padding-top': '2rem'
+      },
+      selector: '[data-visual-test="anchor-blank"]',
+      simulateSelector: '[data-visual-test="anchor-blank"] a',
+      waitAfterSimulate: 200,
+      simulate: 'active'
     })
     expect(screenshot).toMatchImageSnapshot()
   })